requirements.txt
modal==0.56.4434
jupyter==1.0.0

python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
transformers[torch]==4.47.1
evaluate==0.4.1
datasets==2.16.1
scikit-learn==1.3.2
wandb==0.16.1
uvicorn==0.34.0
pytorch-lightning==2.4.0
tensorboard==2.18.0
tensorboardX==2.6.2.2